Created subject controller to separate from admin controller
